Numerical Stabilization & Major Changes
We added the option to add a slack to the optimization problems. This
slack is also set as objective of the linear problem This increases the
stability of the solution. The expression of slack can be set via the
new option 'objective'.
New option 'feasibilityTest' which  determines which if numerical
infeasible options are repelled or not.
If options were not set, they will be cast to be default. Used options
are passed back.
This version works so far only for DSOS based methods. HR will follow
soon.